+++ js/filter-modal.js
@@ -1,6 +1,8 @@
 
 import {body} from './modal.js';
 import {isEscapeKey} from './utils.js';
+import {imgUploadEffectLevel} from './effects.js';
+
 
 const imgUploadInput = document.querySelector('.img-upload__input');
 const imgUploadOverlay = document.querySelector('.img-upload__overlay');
@@ -9,7 +11,6 @@ const imgUploadOverlay = document.querySelector('.img-upload__overlay');
 const imgUploadPreview = document.querySelector('.img-upload__preview img');
 const imgUploadBtnCancel = document.querySelector('.img-upload__cancel');
 
-//const previewCollection = [...document.querySelectorAll('.effects__preview')];
 
 //Нажатие escape
 const onDocumentKeyDown = (evt) => {
@@ -22,19 +23,11 @@ const onDocumentKeyDown = (evt) => {
 
 const allEffects = ['effects__preview--none', 'effects__preview--chrome','effects__preview--sepia','effects__preview--marvin','effects__preview--phobos','effects__preview--heat'];
 
-const [none] = allEffects;
-let effectForClose = '';
-const saveEffectForClose = (effect) => {
-  effectForClose = effect;
-};
 
 //Закрытие окна. Удаление обработчика escape. Удаление обработчика клика,Удаление посл класса с больш фото
 const closeUploadModal = () => {
   imgUploadOverlay.classList.add('hidden');
   body.classList.remove('modal-open');
-  //по умолчанию класс 'effects__preview--none'
-  imgUploadPreview.classList.remove(effectForClose);
-  imgUploadPreview.classList.add(none);
 
   document.removeEventListener('keydown', onDocumentKeyDown);
   imgUploadBtnCancel.removeEventListener('click', closeUploadModal);
@@ -44,7 +37,7 @@ const closeUploadModal = () => {
 const openUploadModal = () => {
   imgUploadOverlay.classList.remove('hidden');
   body.classList.add('modal-open');
-    //imgUploadEffectLevel.classList.add('hidden');
+  imgUploadEffectLevel.classList.add('hidden');
 
   document.addEventListener('keydown', onDocumentKeyDown);
   imgUploadBtnCancel.addEventListener('click', closeUploadModal);
@@ -69,4 +62,4 @@ imgUploadInput.addEventListener('change', (evt) => {
   imgUploadInput.value = '';
 });
 
-export{allEffects,saveEffectForClose,imgUploadPreview};
+export{allEffects,imgUploadPreview};
