+++ js/main.js
@@ -1,108 +1,112 @@
 import {createArraysOfPhotos} from './data.js';
 import {renderPhotos} from './renderThumbnails.js';
 
-import{EFFECTS,getChromeStyleFilter,getSepiaStyleFilter,getMarvinStyleFilter,getPhobosStyleFilter,getHeatStyleFilter} from './constants.js';
+
 const data = createArraysOfPhotos();
 renderPhotos(data);
-import {allEffects,saveEffectForClose,imgUploadPreview} from './filter-modal.js';
-import{scaleInput} from './scale.js'; //Он здесь не нужен, сделан, чтоб подключить соответствующий модуль. Как правильно подключить этот модуль?
-const [none,chrome,sepia,marvin,phobos,heat] = allEffects;
 
-//Вешаю фильтр на фотографию. Эффекты эКСПОРТ imgUploadPreview
-//Ставлю слушатель на ul effects__list
-const effectsList = document.querySelector('.effects__list');
-const effectLevelInput = document.querySelector('.effect-level__value');//ЧИСЛО ДОЛЖНО БЫТЬ
+import {closeSliderByClick} from './effects.js';
 
-effectsList.addEventListener('click', (evt) => {
-  if (evt.target.classList.contains('effects__preview')) {
-    const searchForClass = (effects) => {
+// import{EFFECTS,getChromeStyleFilter,getSepiaStyleFilter,getMarvinStyleFilter,getPhobosStyleFilter,getHeatStyleFilter} from './constants.js';
+// import {allEffects,saveEffectForClose,imgUploadPreview} from './filter-modal.js';
+// import{scaleInput} from './scale.js'; //Он здесь не нужен, сделан, чтоб подключить соответствующий модуль. Как правильно подключить этот модуль?
+// const [none,chrome,sepia,marvin,phobos,heat] = allEffects;
 
-      effects.forEach((effect) => {
-        imgUploadPreview.classList.remove(effect);
-        if(evt.target.classList.contains(effect)) {
-          imgUploadPreview.classList.add(effect);
-          saveEffectForClose(effect);
-          saveEffectForLevel(effect);//вытаскивает спан, т.к. на нем класс хром ит.д
-        }
-      });
-    };
-    searchForClass(allEffects);
-  }
-}
-);
+// //Вешаю фильтр на фотографию. Эффекты эКСПОРТ imgUploadPreview
+// //Ставлю слушатель на ul effects__list
+// const effectsList = document.querySelector('.effects__list');
+// const effectLevelInput = document.querySelector('.effect-level__value');//ЧИСЛО ДОЛЖНО БЫТЬ
 
-//В ХТМЛ:Подключила библиотеку. Подключила СТИЛИ библиотеки - появился белый ползунок
-const sliderContainer = document.querySelector('.effect-level__slider');
-//дефолтное значение спрятанного инпута
-effectLevelInput.value = 100;
+// effectsList.addEventListener('click', (evt) => {
+//   if (evt.target.classList.contains('effects__preview')) {
+//     const searchForClass = (effects) => {
 
-const imgUploadEffectLevel = document.querySelector('.img-upload__effect-level');
-//Создала первичные настройки для слайдера
-noUiSlider.create(sliderContainer, {
-  start: 100,
-  step: 1,
-  connect: 'lower',
-  range: {
-    min: 0,
-    max: 100,
-  },
-  format: {
-    to: (value) => value,
-    from: (value) => parseFloat(value).toFixed(0),
-  },
-});
+//       effects.forEach((effect) => {
+//         imgUploadPreview.classList.remove(effect);
+//         if(evt.target.classList.contains(effect)) {
+//           imgUploadPreview.classList.add(effect);
+//           saveEffectForClose(effect);
+//           saveEffectForLevel(effect);//вытаскивает спан, т.к. на нем класс хром ит.д
+//         }
+//       });
+//     };
+//     searchForClass(allEffects);
+//   }
+// }
+// );
 
-//Ищу все радиокнопки
-const effectRadioInputs = [...document.querySelectorAll('.effects__radio')];
-//По Клику на чекбокс возникает change, добавляю параметры конкретного фильтра
-effectRadioInputs.forEach((radioBtn)=> {
-  radioBtn.addEventListener('change',(evt) => {
-    if(evt.target.checked) {
-      const currentEffect = evt.target.value;
-      //изменяем настройки для конкретного фильтра {range,start и т д}
-    sliderContainer.noUiSlider.updateOptions(EFFECTS[currentEffect]);//VALUE - CHROME
-    console.log(currentEffect)
-    //getStyleToPhoto(currentEffect);
-   // selectEffect(currentEffect);
-   //Ставим слушатель обновления события на слайдер
-      sliderContainer.noUiSlider.on('update',() => {
-  //запись в спрятанный инпут значение ползунка
-        effectLevelInput.value = sliderContainer.noUiSlider.get();
-        console.log(effectLevelInput.value);
-        switch(currentEffect){
-          case 'chrome' :
-      imgUploadPreview.style.filter = getChromeStyleFilter(effectLevelInput.value);
-      break;
-      case 'sepia' :
-        imgUploadPreview.style.filter = getSepiaStyleFilter(effectLevelInput.value);
-        break;
-        case 'marvin' :
-          imgUploadPreview.style.filter = getMarvinStyleFilter(effectLevelInput.value);
-          break;
-          case 'phobos' :
-            imgUploadPreview.style.filter = getPhobosStyleFilter(effectLevelInput.value);
-            break;
-            case 'heat' :
-            imgUploadPreview.style.filter = getHeatStyleFilter(effectLevelInput.value);
-            break;
-            default:
-              imgUploadPreview.style.filter = 'none';
-              }
-        });
-    }
-  });
-});
+// //В ХТМЛ:Подключила библиотеку. Подключила СТИЛИ библиотеки - появился белый ползунок
+// const sliderContainer = document.querySelector('.effect-level__slider');
+// //дефолтное значение спрятанного инпута
+// effectLevelInput.value = 100;
+
+// const imgUploadEffectLevel = document.querySelector('.img-upload__effect-level');
+// //Создала первичные настройки для слайдера
+// noUiSlider.create(sliderContainer, {
+//   start: 100,
+//   step: 1,
+//   connect: 'lower',
+//   range: {
+//     min: 0,
+//     max: 100,
+//   },
+//   format: {
+//     to: (value) => value,
+//     from: (value) => parseFloat(value).toFixed(0),
+//   },
+// });
+
+// //Ищу все радиокнопки
+// const effectRadioInputs = [...document.querySelectorAll('.effects__radio')];
+// //По Клику на чекбокс возникает change, добавляю параметры конкретного фильтра
+// effectRadioInputs.forEach((radioBtn)=> {
+//   radioBtn.addEventListener('change',(evt) => {
+//     if(evt.target.checked) {
+//       const currentEffect = evt.target.value;
+//       //изменяем настройки для конкретного фильтра {range,start и т д}
+//     sliderContainer.noUiSlider.updateOptions(EFFECTS[currentEffect]);//VALUE - CHROME
+//     console.log(currentEffect)
+//     //getStyleToPhoto(currentEffect);
+//    // selectEffect(currentEffect);
+//    //Ставим слушатель обновления события на слайдер
+//       sliderContainer.noUiSlider.on('update',() => {
+//   //запись в спрятанный инпут значение ползунка
+//         effectLevelInput.value = sliderContainer.noUiSlider.get();
+//         console.log(effectLevelInput.value);
+//         switch(currentEffect){
+//           case 'chrome' :
+//       imgUploadPreview.style.filter = getChromeStyleFilter(effectLevelInput.value);
+//       break;
+//       case 'sepia' :
+//         imgUploadPreview.style.filter = getSepiaStyleFilter(effectLevelInput.value);
+//         break;
+//         case 'marvin' :
+//           imgUploadPreview.style.filter = getMarvinStyleFilter(effectLevelInput.value);
+//           break;
+//           case 'phobos' :
+//             imgUploadPreview.style.filter = getPhobosStyleFilter(effectLevelInput.value);
+//             break;
+//             case 'heat' :
+//             imgUploadPreview.style.filter = getHeatStyleFilter(effectLevelInput.value);
+//             break;
+//             default:
+//               imgUploadPreview.style.filter = 'none';
+//               }
+//         });
+//     }
+//   });
+// });
 
 
-const saveEffectForLevel = (effect) => {
-  if (effect === none) {
-    imgUploadEffectLevel.classList.add('hidden');
-    //imgUploadPreview.style.filter = 'none';
-  } else {
-    imgUploadEffectLevel.classList.remove('hidden');
-  }
-};
-
+// const saveEffectForLevel = (effect) => {
+//   if (effect === none) {
+//     imgUploadEffectLevel.classList.add('hidden');
+//     //imgUploadPreview.style.filter = 'none';
+//   } else {
+//     imgUploadEffectLevel.classList.remove('hidden');
+//   }
+// };
+///////////////////////////////////////////////////
 
 //export {imgUploadEffectLevel};
